+++ craft/app/services/AssetTransformsService.php
@@ -32,6 +32,16 @@ class AssetTransformsService extends BaseApplicationComponent
 	 */
 	private $_sourcesToBeDeleted = array();
 
+	/**
+	 * @var array
+	 */
+	private $_eagerLoadedTransformIndexes;
+
+	/**
+	 * @var
+	 */
+	private $_activeTransformIndexModel;
+
 	// Public Methods
 	// =========================================================================
 
@@ -196,6 +206,108 @@ class AssetTransformsService extends BaseApplicationComponent
 	}
 
 	/**
+	 * Eager-loads transform indexes for a given set of file IDs.
+	 *
+	 * @param AssetFileModel[] $files The files to eager-load tranforms for
+	 * @param array $transforms The transform definitions to eager-load
+	 *
+	 * @return void
+	 */
+	public function eagerLoadTransforms($files, $transforms)
+	{
+		if (!$files || !$transforms)
+		{
+			return;
+		}
+
+		// Index the files by ID
+		$filesById = array();
+		foreach ($files as $file)
+		{
+			$filesById[$file->id] = $file;
+		}
+
+		// Get the index conditions
+		$transformsByFingerprint = array();
+		$indexConditions = array();
+
+		foreach ($transforms as $transform)
+		{
+			$transform = $this->normalizeTransform($transform);
+			$location = $fingerprint = $this->_getTransformFolderName($transform);
+
+			$condition = array('and', array('location' => $location));
+
+			if (is_null($transform->format))
+			{
+				$condition[] = 'format IS NULL';
+			}
+			else
+			{
+				$condition[] = array('format' => $transform->format);
+				$fingerprint .= ':'.$transform->format;
+			}
+
+			$indexConditions[] = $condition;
+			$transformsByFingerprint[$fingerprint] = $transform;
+		}
+
+		if (count($indexConditions) == 1)
+		{
+			$indexConditions = $indexConditions[0];
+		}
+		else
+		{
+			array_unshift($indexConditions, 'or');
+		}
+
+		// Query for the indexes
+		$results = craft()->db->createCommand()
+			->select('*')
+			->from('assettransformindex')
+			->where(array('and',
+				array('in', 'fileId', array_keys($filesById)),
+				$indexConditions
+			))
+			->queryAll();
+
+		// Index the valid transform indexes by fingerprint, and capture the IDs of indexes that should be deleted
+		$resultsByFingerprint = array();
+		$invalidIndexIds = array();
+
+		foreach ($results as $result)
+		{
+			// Get the transform's fingerprint
+			$transformFingerprint = $result['location'];
+
+			if ($result['format'])
+			{
+				$transformFingerprint .= ':'.$result['format'];
+			}
+
+			// Is it still valid?
+			$transform = $transformsByFingerprint[$transformFingerprint];
+			$file = $filesById[$result['fileId']];
+
+			if ($this->validateTransformIndexResult($result, $transform, $file))
+			{
+				$indexFingerprint = $result['fileId'].':'.$transformFingerprint;
+				$this->_eagerLoadedTransformIndexes[$indexFingerprint] = $result;
+			}
+			else
+			{
+				$invalidIndexIds[] = $result['id'];
+			}
+		}
+
+		// Delete any invalid indexes
+		if ($invalidIndexIds)
+		{
+			craft()->db->createCommand()->delete('assettransformindex', array('in', 'id', $invalidIndexIds));
+		}
+	}
+
+	/**
 	 * Get a transform index row. If it doesn't exist - create one.
 	 *
 	 * @param AssetFileModel $file
@@ -207,6 +319,16 @@ class AssetTransformsService extends BaseApplicationComponent
 	{
 		$transform = $this->normalizeTransform($transform);
 		$transformLocation = $this->_getTransformFolderName($transform);
+
+		// Was it eager-loaded?
+		$fingerprint = $file->id.':'.$transformLocation.(is_null($transform->format) ? '' : ':'.$transform->format);
+
+		if (isset($this->_eagerLoadedTransformIndexes[$fingerprint]))
+		{
+			$entry = $this->_eagerLoadedTransformIndexes[$fingerprint];
+
+			return new AssetTransformIndexModel($entry);
+		}
 
 		// Check if an entry exists already
 		$query = craft()->db->createCommand()
@@ -229,14 +351,7 @@ class AssetTransformsService extends BaseApplicationComponent
 
 		if ($entry)
 		{
-			// If the file has been indexed after any changes impacting the transform, return the record
-			$indexedAfterFileModified = $entry['dateIndexed'] >= $file->dateModified->format(DateTime::MYSQL_DATETIME, DateTime::UTC);
-			$indexedAfterTransformParameterChange =
-				(!$transform->isNamedTransform()
-					|| ($transform->isNamedTransform()
-						&& $entry['dateIndexed'] >= $transform->dimensionChangeTime->format(DateTime::MYSQL_DATETIME, DateTime::UTC)));
-
-			if ($indexedAfterFileModified && $indexedAfterTransformParameterChange)
+			if ($this->validateTransformIndexResult($entry, $transform, $file))
 			{
 				return new AssetTransformIndexModel($entry);
 			}
@@ -265,6 +380,33 @@ class AssetTransformsService extends BaseApplicationComponent
 	}
 
 	/**
+	 * Validates a transform index result to see if the index is still valid for a given file.
+	 *
+	 * @param array               $result
+	 * @param AssetTransformModel $transform
+	 * @param AssetFileModel      $file
+	 *
+	 * @return bool Whether the index result is still valid
+	 */
+	public function validateTransformIndexResult($result, AssetTransformModel $transform, AssetFileModel $file)
+	{
+		$indexedAfterFileModified = $result['dateIndexed'] >= $file->dateModified->format(DateTime::MYSQL_DATETIME, DateTime::UTC);
+		$indexedAfterTransformParameterChange =
+			(!$transform->isNamedTransform()
+				|| ($transform->isNamedTransform()
+					&& $result['dateIndexed'] >= $transform->dimensionChangeTime->format(DateTime::MYSQL_DATETIME, DateTime::UTC)));
+
+		if ($indexedAfterFileModified && $indexedAfterTransformParameterChange)
+		{
+			return true;
+		}
+		else
+		{
+			return false;
+		}
+	}
+
+	/**
 	 * Get a transform URL by the transform index model.
 	 *
 	 * @param AssetTransformIndexModel $index
@@ -281,7 +423,7 @@ class AssetTransformsService extends BaseApplicationComponent
 			{
 				// Wait a second!
 				sleep(1);
-				ini_set('max_execution_time', 120);
+				craft()->config->maxPowerCaptain();
 
 				$index = $this->getTransformIndexModelById($index->id);
 
@@ -297,7 +439,6 @@ class AssetTransformsService extends BaseApplicationComponent
 					}
 					else
 					{
-						$index->dateUpdated = new DateTime();
 						$this->storeTransformIndexData($index);
 						break;
 					}
@@ -411,14 +552,14 @@ class AssetTransformsService extends BaseApplicationComponent
 		// If we have a match, copy the file.
 		if ($matchFound)
 		{
-			$source->copyTransform($file, $file->getFolder(), new AssetTransformIndexModel($matchFound), $index);
+			$source->copyTransform($file, $file->folderPath, new AssetTransformIndexModel($matchFound), $index);
 		}
 		else
 		{
 			$this->_createTransformForFile($file, $index);
 		}
 
-		return $source->fileExists($file->getFolder()->path.$this->getTransformSubfolder($file, $index), $this->getTransformFilename($file, $index));
+		return $source->fileExists($file->folderPath.$this->getTransformSubfolder($file, $index), $this->getTransformFilename($file, $index));
 	}
 
 	/**
@@ -580,11 +721,25 @@ class AssetTransformsService extends BaseApplicationComponent
 	public function getUrlForTransformByTransformIndex(AssetTransformIndexModel $transformIndexModel)
 	{
 		$file = craft()->assets->getFileById($transformIndexModel->fileId);
+
+		return $this->getUrlForTransformByAssetAndTransformIndex($file, $transformIndexModel);
+	}
+
+	/**
+	 * Get the URL for transform for the given asset, by the given transform index model
+	 *
+	 * @param AssetFileModel           $file
+	 * @param AssetTransformIndexModel $transformIndexModel
+	 *
+	 * @return string
+	 */
+	public function getUrlForTransformByAssetAndTransformIndex(AssetFileModel $file, AssetTransformIndexModel $transformIndexModel)
+	{
 		$sourceType = craft()->assetSources->getSourceTypeById($file->sourceId);
 		$baseUrl = $sourceType->getBaseUrl();
 		$appendix = AssetsHelper::getUrlAppendix($sourceType, $file);
 
-		return $baseUrl . $file->getFolder()->path . $this->getTransformSubpath($file, $transformIndexModel) . $appendix;
+		return $baseUrl . $file->folderPath . $this->getTransformSubpath($file, $transformIndexModel) . $appendix;
 	}
 
 	/**
@@ -663,7 +818,7 @@ class AssetTransformsService extends BaseApplicationComponent
 		{
 			if (!$sourceType->isRemote())
 			{
-				throw new Exception(Craft::t('Image “{file}” cannot be found.', array('file' => $file->filename)));
+				throw new Exception(Craft::t('Image “{file}” cannot be found.', array('file' => $imageSourcePath)));
 			}
 
 			// Delete it just in case it's a 0-byter
@@ -899,11 +1054,14 @@ class AssetTransformsService extends BaseApplicationComponent
 	{
 		$thumbFolders = IOHelper::getFolderContents(craft()->path->getAssetsThumbsPath());
 
-		foreach ($thumbFolders as $folder)
+		if ($thumbFolders)
 		{
-			if (is_dir($folder))
+			foreach ($thumbFolders as $folder)
 			{
-				IOHelper::deleteFile($folder.'/'.$file->id.'.'.$this->_getThumbExtension($file));
+				if (is_dir($folder))
+				{
+					IOHelper::deleteFile($folder.'/'.$file->id.'.'.$this->_getThumbExtension($file));
+				}
 			}
 		}
 	}
@@ -941,6 +1099,22 @@ class AssetTransformsService extends BaseApplicationComponent
 			->queryAll();
 
 		return AssetTransformIndexModel::populateModels($records);
+	}
+
+	/**
+	 * @return mixed
+	 */
+	public function getActiveTransformIndexModel()
+	{
+		return $this->_activeTransformIndexModel;
+	}
+
+	/**
+	 * @param AssetTransformIndexModel $index
+	 */
+	public function setActiveTransformIndexModel(AssetTransformIndexModel $index)
+	{
+		$this->_activeTransformIndexModel = $index;
 	}
 
 	// Private Methods
@@ -1057,6 +1231,9 @@ class AssetTransformsService extends BaseApplicationComponent
 		{
 			$image->setQuality($quality);
 		}
+
+		// Save this for Image to use if needed.
+		$this->setActiveTransformIndexModel($index);
 
 		switch ($transform->mode)
 		{
